Shell
Creates a textual shell that simulates bash in Linux and has a few different commands. "exit" will terminate the shell. "SERIAL" followed by a second word--file name containing commands to be run--will run those commands serially. "PARALLEL" will do the same but run the commands in parallel. If not one of the above, it is assumed to be a program name and any command line arguments for that program.